requirements.txt
appdirs==1.4.4
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.5
django-leaflet==0.28.0
djangorestframework==3.12.4
filelock==3.0.12
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
mccabe==0.6.1
Pillow==8.3.1
pipenv==2021.5.29
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tinydb==4.4.0
urllib3==1.26.6
virtualenv==20.4.7
virtualenv-clone==0.5.4
whitenoise==5.2.0
